Move EnumUtility to common
